+++ apps/api/serializers/feeds.py
@@ -1,9 +1,11 @@
 from datetime import datetime
-from typing import List
+from typing import List, Optional
 from uuid import UUID
 
-from porcupine.base import Serializer
+from pydantic import field_validator, Field
+from pydantic_core.core_schema import ValidationInfo
 
+from apps.api.serializers import Serializer
 from apps.core.models import Feed
 
 
@@ -11,22 +13,22 @@ class FeedSerializer:
     class Base(Serializer):
         id: UUID
         catalog_id: UUID
-        parents: list = []
-        children: list = []
+        parents: List[UUID] = Field(default=[], validate_default=True)
+        children: List[UUID] = Field(default=[], validate_default=True)
         creator_id: UUID
         title: str
         kind: Feed.FeedKind
         url_name: str
         url: str
         content: str
-        per_page: int = None
+        per_page: Optional[int]
         created_at: datetime
         updated_at: datetime
 
-        @staticmethod
-        def resolve_parents(data: Feed, **kwargs) -> List[UUID]:
-            return list(data.parents.all().values_list('id', flat=True))
+        @field_validator('parents', mode='before')
+        def generate_parents(cls, v, info: ValidationInfo) -> List[UUID]:
+            return v.all().values_list('id', flat=True)
 
-        @staticmethod
-        def resolve_children(data: Feed, **kwargs) -> List[UUID]:
-            return list(data.children.all().values_list('id', flat=True))
+        @field_validator('children', mode='before')
+        def generate_children(cls, v, info: ValidationInfo) -> List[UUID]:
+            return v.all().values_list('id', flat=True)